performance_tests/mongo_postgres/requirements.txt
pydantic==1.10.11
pymongo==4.5.0
psycopg2-binary==2.9.5